server/requirements.txt
anchorpy==0.21.0
anchorpy-core==0.2.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
base58==2.1.1
based58==0.1.1
borsh-construct==0.1.0
boto3==1.40.36
botocore==1.40.36
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
Django==5.2.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-silk==5.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gprof2dot==2025.4.14
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonalias==0.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyheck==0.1.5
PyJWT==2.10.1
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
solana==0.36.6
solders==0.26.0
sqlparse==0.5.3
starlette==0.48.0
sumtypes==0.1a6
toml==0.10.2
toolz==0.11.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0
whitenoise==6.11.0
